Fix check for trailing garbage in basebackup import.
There was a warning for trailing garbage after end-of-tar archive, but it
didn't always work. The reason is that we created a StreamReader over
the original copyin-stream, but performed the check for garbage on the
copyin-stream. There could be some garbage bytes buffered in the
StreamReader, which were not caught by the warning.

I considered turning the the warning into a fatal error, aborting the
import, but I wasn't sure if we handle aborting the import properly.
Do we clean up the timeline directory on error? If we don't, we should
make that more robust, but that's a different story.